refactor : image path key generate 모듈 변경
