+++ src/main/java/univ/earthbreaker/namu/core/api/mission/MissionPostImagePathKeyGenerator.java
@@ -1,4 +1,4 @@
-package univ.earthbreaker.namu.external.aws.image;
+package univ.earthbreaker.namu.core.api.mission;
 
 import java.util.UUID;
 
@@ -6,23 +6,25 @@ import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-@Component
-public class PostImagePathKeyGenerator implements ImagePathKeyGenerator {
+import univ.earthbreaker.namu.external.aws.image.ImagePathKeyGenerator;
+
+@Component(value = "missionPostImagePathGen")
+public class MissionPostImagePathKeyGenerator implements ImagePathKeyGenerator {
 
 	private static final String POST_IMAGE_PATH_KEY_FORMAT = "%s/%s_%s_%s";
 
 	private final String postImageDir;
 
-	public PostImagePathKeyGenerator(@Value("${image.post-dir}") String postImageDir) {
+	public MissionPostImagePathKeyGenerator(@Value("${image.post-dir}") String postImageDir) {
 		this.postImageDir = postImageDir;
 	}
 
 	@Override
-	public @NotNull String generate(long memberNo, String originFileName) {
+	public @NotNull String generate(String memberKey, String originFileName) {
 		return String.format(
 			POST_IMAGE_PATH_KEY_FORMAT,
 			postImageDir,
-			memberNo,
+			memberKey,
 			UUID.randomUUID(),
 			originFileName
 		);
