+++ src/main/java/univ/earthbreaker/namu/external/aws/image/ImagePathKeyGenerator.java
@@ -3,5 +3,5 @@ package univ.earthbreaker.namu.external.aws.image;
 import org.jetbrains.annotations.NotNull;
 
 public interface ImagePathKeyGenerator {
-	@NotNull String generate(long memberNo, String originFileName);
+	@NotNull String generate(String memberKey, String originFileName);
 }
