+++ src/main/resources/static/docs/mission-api.html
@@ -1030,7 +1030,7 @@ Host: namu.api.docs</code></pre>
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2024-04-01 18:37:00 +0900
+Last updated 2024-04-04 17:55:39 +0900
 </div>
 </div>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.18.3/highlight.min.js"></script>
